+++ helpers/__tests__/HTMLView.test.js
@@ -1,7 +1,8 @@
 /* global describe, expect, test, beforeAll */
 
 import colors from 'chalk'
-import * as n from '../HTMLView'
+import * as h from '../HTMLView'
+import * as n from '../NPThemeToCSS'
 import { Calendar, Clipboard, CommandBar, DataStore, Editor, NotePlan, Note, Paragraph } from '@mocks/index'
 
 beforeAll(() => {
@@ -24,45 +25,45 @@ describe(`${FILE}`, () => {
    */
   describe('generateScriptTags()' /* function */, () => {
     test('should return empty if scripts is undefined', () => {
-      const result = n.generateScriptTags(undefined)
+      const result = h.generateScriptTags(undefined)
       expect(result).toEqual(``)
     })
     test('should return empty if scripts is null', () => {
-      const result = n.generateScriptTags(null)
+      const result = h.generateScriptTags(null)
       expect(result).toEqual(``)
     })
     test('should return empty if scripts is empty string', () => {
-      const result = n.generateScriptTags('')
+      const result = h.generateScriptTags('')
       expect(result).toEqual(``)
     })
     test('should not add <script> tag if STRING already has it', () => {
       const input = '<script>foo</script>'
-      const result = n.generateScriptTags(input)
+      const result = h.generateScriptTags(input)
       expect(result).toEqual(`${input}\n`)
     })
     test('should add <script> tag if STRING does not have it', () => {
       const input = 'foo'
-      const result = n.generateScriptTags(input)
+      const result = h.generateScriptTags(input)
       expect(result).toEqual(`<script type="text/javascript">\n${input}\n</script>\n`)
     })
     test('should add <script> tag if OBJ does not have it', () => {
       const input = { code: 'foo' }
-      const result = n.generateScriptTags(input)
+      const result = h.generateScriptTags(input)
       expect(result).toEqual(`<script type="text/javascript">\nfoo\n</script>\n`)
     })
     test('should not add <script> tag if OBJ does have it', () => {
       const input = { code: '<script>foo</script>' }
-      const result = n.generateScriptTags(input)
+      const result = h.generateScriptTags(input)
       expect(result).toEqual('<script>foo</script>\n')
     })
     test('should add <script type="xxx"> tag if OBJ does have it', () => {
       const input = { code: 'foo', type: 'bar' }
-      const result = n.generateScriptTags(input)
+      const result = h.generateScriptTags(input)
       expect(result).toEqual('<script type="bar">\nfoo\n</script>\n')
     })
     test('should add multiple mixed types', () => {
       const input = [{ code: 'foo', type: 'bar' }, 'foo']
-      const result = n.generateScriptTags(input)
+      const result = h.generateScriptTags(input)
       expect(result).toEqual(`<script type="bar">\nfoo\n</script>\n\n<script type="text/javascript">\nfoo\n</script>\n`)
     })
   })
@@ -74,25 +75,25 @@ describe(`${FILE}`, () => {
     test('should do nothing if nothing to trim (empty)', () => {
       const orig = {}
       const expected = null
-      const result = n.pruneTheme(orig)
+      const result = h.pruneTheme(orig)
       expect(result).toEqual(expected)
     })
     test('should do nothing if nothing to trim 2', () => {
       const orig = { foo: 'bar' }
       const expected = {}
-      const result = n.pruneTheme(orig)
+      const result = h.pruneTheme(orig)
       expect(result).toEqual(orig)
     })
     test('should trim a top level item', () => {
       const orig = { __orderedStyles: ['title-mark1', 'title-mark2'] }
       const expected = null
-      const result = n.pruneTheme(orig)
+      const result = h.pruneTheme(orig)
       expect(result).toEqual(expected)
     })
     test('should trim a top level item but leave others', () => {
       const orig = { __orderedStyles: ['title-mark1', 'title-mark2'], name: 'foo' }
       const expected = { name: 'foo' }
-      const result = n.pruneTheme(orig)
+      const result = h.pruneTheme(orig)
       expect(result).toEqual(expected)
     })
     test('should remove properties inside of a style', () => {
@@ -116,7 +117,7 @@ describe(`${FILE}`, () => {
           },
         },
       }
-      const result = n.pruneTheme(orig)
+      const result = h.pruneTheme(orig)
       expect(result).toEqual(expected)
     })
     test('should remove property that has empty value', () => {
@@ -141,7 +142,7 @@ describe(`${FILE}`, () => {
           },
         },
       }
-      const result = n.pruneTheme(orig)
+      const result = h.pruneTheme(orig)
       expect(result).toEqual(expected)
     })
     test('should remove properties inside of a style', () => {
@@ -166,7 +167,7 @@ describe(`${FILE}`, () => {
           },
         },
       }
-      const result = n.pruneTheme(orig)
+      const result = h.pruneTheme(orig)
       expect(result).toEqual(expected)
     })
     test('should remove empty objects after removing all props', () => {
@@ -181,7 +182,7 @@ describe(`${FILE}`, () => {
         },
       }
       const expected = null
-      const result = n.pruneTheme(orig)
+      const result = h.pruneTheme(orig)
       expect(result).toEqual(expected)
     })
   })
@@ -193,18 +194,18 @@ describe(`${FILE}`, () => {
 describe('replaceMarkdownLinkWithHTMLLink()' /* function */, () => {
   test('should not do anything if no url', () => {
     const orig = 'foo bar'
-    const result = n.replaceMarkdownLinkWithHTMLLink(orig)
+    const result = h.replaceMarkdownLinkWithHTMLLink(orig)
     expect(result).toEqual(orig)
   })
   test('should replace a url', () => {
     const orig = 'foo [link](http://) bar'
-    const result = n.replaceMarkdownLinkWithHTMLLink(orig)
+    const result = h.replaceMarkdownLinkWithHTMLLink(orig)
     const expected = 'foo <a href="http://">link</a> bar'
     expect(result).toEqual(expected)
   })
   test('should replace > 1 url', () => {
     const orig = 'foo [link](http://) bar [link2](http://) baz [link3](noteplan://)'
-    const result = n.replaceMarkdownLinkWithHTMLLink(orig)
+    const result = h.replaceMarkdownLinkWithHTMLLink(orig)
     const expected = 'foo <a href="http://">link</a> bar <a href="http://">link2</a> baz <a href="noteplan://">link3</a>'
     expect(result).toEqual(expected)
   })