+++ jgclark.WindowSets/src/windowSets.js
@@ -2,22 +2,30 @@
 //---------------------------------------------------------------
 // Main functions for WindowSets plugin
 // Jonathan Clark
-// last update 29.8.2022 for v0.2.0 by @jgclark
+// last update 10.9.2022 for v0.2.0 by @jgclark
 //---------------------------------------------------------------
 
 import pluginJson from '../plugin.json'
-import { getCodeBlocksOfType } from '@helpers/codeBlocks'
+import { addCodeBlock, getCodeBlocksOfType, updateCodeBlock } from '@helpers/codeBlocks'
 import {
   calcOffsetDateStr,
+  daysBetween,
+  // getDateStringFromCalendarFilename,
+  // getDateStrForStartofPeriodFromCalendarFilename,
+  getDateStringFromCalendarFilename,
+  getISODateStringFromYYYYMMDD,
   getTodaysDateHyphenated,
   isValidCalendarNoteFilename,
   RE_OFFSET_DATE_CAPTURE,
-  RE_OFFSET_DATE
+  RE_OFFSET_DATE,
+  unhyphenateString,
+  // unhyphenatedDate,
 } from '@helpers/dateTime'
 import { clo, JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
 import { logPreference, unsetPreference } from '@helpers/NPdev'
 import { displayTitle } from '@helpers/general'
-import { getProjectNotesInFolder } from '@helpers/note'
+import { getOrMakeNote, getProjectNotesInFolder } from '@helpers/note'
+import { relativeDateCodeFromDateString } from '@helpers/NPdateTime'
 import { getNoteFilenameFromTitle } from '@helpers/NPnote'
 import {
   applyRectToWindow,
@@ -32,14 +40,24 @@ import {
 import { chooseOption, getInputTrimmed, showMessage, showMessageYesNo, showMessageYesNoCancel } from '@helpers/userInput'
 
 //-----------------------------------------------------------------
+// Plugin lookup list
+type PluginWindowCommand = {
+  pluginWindowId: string,
+  pluginID: string,
+  pluginCommandName: string
+}
+const pluginWindowsAndCommands: Array<PluginWindowCommand> = [
+  { pluginWindowId: 'jgclark.Dashboard.main', pluginID: 'jgclark.Dashboard', pluginCommandName: 'Show Dashboard window' },
+  { pluginWindowId: 'Review List', pluginID: 'jgclark.Reviews', pluginCommandName: 'project lists' },
+]
+
+//-----------------------------------------------------------------
 // Data types
 type EditorWinDetails = {
-  // id?: string, // TEST: confirm this can be lost for definitions
-  type: string, // make mandatory again in time. "Calendar" | "Note"
+  noteType: string, // "Calendar" | "Note"
   filename: string,
-  openAction: string, // "floating" or "split"
-  customId?: string, // make mandatory again in time
-  title?: string,
+  openAction: string, // "main" | "floating" | "split"
+  title?: string, // optional, but persist it where used
   x?: number,
   y?: number,
   width?: number,
@@ -84,7 +102,7 @@ This is an example of:
 - use of relative dates to define dates that work relative to today's date (see the [Plugin's README](https://github.com/NotePlan/plugins/blob/main/jgclark.WindowSets/README.md) for more details)
 - the ability to "closeOtherWindows" before the Window Set is activated
 - "htmlWindows", "x", "y", "width", "height" should be left out if you don't want to define them
-- these can have customIDs, though currently they aren't used.
+- these can have "title"s, though currently they're only used to help you identify different windows.
 
 Note: the JSON has to be well-formatted to be usable. In particular check that there aren't any extra commas after the final item of any section.
 
@@ -94,22 +112,22 @@ Note: the JSON has to be well-formatted to be usable. In particular check that t
   "closeOtherWindows": true,
   "editorWindows": [
 		{
-		"type": "Calendar",
+		"noteType": "Calendar",
 		"openAction": "split",
 		"filename": "{-1d}",
-		"customID": "yesterday"
+		"title": "yesterday"
 		},
 		{
-		"type": "Calendar",
+		"noteType": "Calendar",
 		"openAction": "split",
 		"filename": "{0d}",
-		"customID": "today"
+		"title": "today"
 		},
 		{
-		"type": "Calendar",
+		"noteType": "Calendar",
 		"openAction": "split",
 		"filename": "{+1d}",
-		"customID": "tomorrow"
+		"title": "tomorrow"
 		}
   ]
 }
@@ -121,22 +139,22 @@ Note: the JSON has to be well-formatted to be usable. In particular check that t
   "closeOtherWindows": true,
   "editorWindows": [
 		{
-		"type": "Calendar",
+		"noteType": "Calendar",
 		"openAction": "split",
 		"filename": "{-1w}",
-		"customID": "last week"
+		"title": "last week"
 		},
 		{
-		"type": "Calendar",
+		"noteType": "Calendar",
 		"openAction": "split",
 		"filename": "{0w}",
-		"customID": "this week"
+		"title": "this week"
 		},
 		{
-		"type": "Calendar",
+		"noteType": "Calendar",
 		"openAction": "split",
 		"filename": "{+1w}",
-		"customID": "next week"
+		"title": "next week"
 		}
   ]
 }
@@ -162,13 +180,13 @@ Note: the JSON has to be well-formatted to be usable. In particular check that t
       "type": "Note",
       "openAction": "floating",
       "filename": "CCC Areas/Admin (Work).md",
-      "customID": "Admin (Work)"
+      "title": "Admin (Work)"
     },
     {
       "type": "Note",
       "openAction": "floating",
       "filename": "Saved Searches/RP Search Results.md",
-      "customID": "@RP Search Results"
+      "title": "@RP Search Results"
     }
   ],
   "htmlWindows": [
@@ -217,14 +235,14 @@ export async function getPluginSettings(): Promise<any> {
 }
 
 /**
- * Read current WindowSets definitions
+ * Read current WindowSet definitions
  * V2: read from 'folderForDefinitions' folder of files. In each the JSON code block(s) define(s) the WS.
  * @param {string} folderForDefinitions
  * @returns {Array<WindowSet>} JSON configuration object for all window sets
  */
 export function readWindowSetDefinitions(folderForDefinitions: string): Array<WindowSet> {
-  logDebug(pluginJson, `readWindowSetDefinitions() starting`)
   try {
+    logDebug(pluginJson, `readWindowSetDefinitions() starting for folder '${folderForDefinitions}'`)
     let windowSets: Array<WindowSet> = []
     // Read all files in @WindowSets folder
     const folderFiles = getProjectNotesInFolder(folderForDefinitions)
@@ -234,7 +252,7 @@ export function readWindowSetDefinitions(folderForDefinitions: string): Array<Wi
         // Get the contents of code blocks ot type JSON
         const codeBlocks = getCodeBlocksOfType(file, ['json'])
         if (codeBlocks.length === 0) {
-          logError('readWindowSetDefinitions', `No JSON code blocks found in '${file.filename}'`)
+          logWarn('readWindowSetDefinitions', `No JSON code blocks found in '${file.filename}'`)
           continue
         } else {
           // Process the codeblock(s)
@@ -251,13 +269,15 @@ export function readWindowSetDefinitions(folderForDefinitions: string): Array<Wi
               editorWindows: thisJSON.editorWindows ?? [],
               htmlWindows: thisJSON.htmlWindows ?? [],
             }
-            clo(thisWS, `thisWS:`)
+            // clo(thisWS, `thisWS:`)
             windowSets.push(thisWS)
           }
         }
       }
+      logDebug('readWindowSetDefinitions', `Read ${String(windowSets.length)} window sets`)
+    } else {
+      logInfo('readWindowSetDefinitions', `No Window Set definition files found in folder '${folderForDefinitions}'`)
     }
-    logDebug('readWindowSetDefinitions', `Read ${String(windowSets.length)} window sets`)
     return windowSets
   } catch (err) {
     logError(pluginJson, `${err.name}: ${err.message}`)
@@ -322,22 +342,21 @@ export async function logWindowSets(): Promise<void> {
       return
     }
     const outputLines = []
-    let c = 0
     outputLines.push(`Window Sets:`)
     for (const set of windowSets) {
-      // clo(set, 'WS #' + String(c))
+      let c = 0
       outputLines.push(`${set.name}:`)
       if (set.editorWindows) {
-        for (const win of set.editorWindows) {
-          outputLines.push(`- ${String(c)}: ${win.type}: customId:'${win.customId ?? ''}' title:'${win.title ?? ''}' filename:${win.filename ?? ''} ID:${win.id} x:${win.x} y:${win.y} w:${win.width} h:${win.height}`)
+        for (const ws of set.editorWindows) {
+          outputLines.push(`- ${String(c)}: ${ws.noteType}: title:'${ws.title ?? ''}' filename:${ws.filename ?? ''} x:${ws.x ?? '-'} y:${ws.y ?? '-'} w:${ws.width ?? '-'} h:${ws.height ?? '-'}`)
           c++
         }
       } else {
         logInfo('logWindowSets', `WindowSet #${String(c)} has no editorWindows array`)
       }
       if (set.htmlWindows) {
-        for (const win of set.htmlWindows) {
-          outputLines.push(`- ${String(c)}: ${win.type}: customId:'${win.customId ?? ''}' title:'${win.title ?? ''}' filename:${win.filename ?? ''} ID:${win.id} x:${win.x} y:${win.y} w:${win.width} h:${win.height}`)
+        for (const ws of set.htmlWindows) {
+          outputLines.push(`- ${String(c)}: ${ws.type}: customId:'${ws.customId ?? ''}' plugin.id:${ws.pluginID ?? '?'} x:${ws.x ?? '-'} y:${ws.y ?? '-'} w:${ws.width ?? '-'} h:${ws.height ?? '-'}`)
           c++
         }
       } else {
@@ -354,7 +373,7 @@ export async function logWindowSets(): Promise<void> {
 
 /**
  * Save detailed set of windows/panes as a set to the preference store for the current device.
- * V1: reads from DataStore.preference('windowSets')
+ * V2: reads from @WindowSets folder of files.
  * @author @jgclark
  */
 export async function saveWindowSet(): Promise<void> {
@@ -365,45 +384,47 @@ export async function saveWindowSet(): Promise<void> {
       logInfo('saveWindowSet', `Window Sets needs NotePlan v3.9.1 or later on macOS. Stopping.`)
       return
     }
-    // Form this set.
+    // Form this set from open windows
     // Note: needs to use a cut-down set of attributes available in the window objects
     const editorWinDetails: Array<EditorWinDetails> = NotePlan.editors.map((win) => {
       const winRect = win.windowRect
       return {
-        id: win.id,
-        customId: win.customId,
-        // title: win.title,
-        type: win.type,
+        noteType: win.type,
+        openAction: '', // set later
         filename: win.filename,
+        title: undefined, // potentially gets defined later
         x: winRect.x,
         y: winRect.y,
         width: winRect.width,
         height: winRect.height,
       }
     })
-    if (editorWinDetails.length < 2) {
+    // logDebug('saveWindowSet', `Found ${String(editorWinDetails.length)} editor windows`)
+
+    const htmlWinDetails: Array<HTMLWinDetails> = NotePlan.htmlWindows.map((win) => {
+      const winRect = win.windowRect
+      return {
+        type: win.type,
+        pluginID: '???', // gets set later
+        pluginCommand: '???', // gets set later
+        customId: win.customId,
+        x: winRect.x,
+        y: winRect.y,
+        width: winRect.width,
+        height: winRect.height,
+      }
+    })
+    // logDebug('saveWindowSet', `Found ${String(htmlWinDetails.length)} HTML windows`)
+
+    if ((editorWinDetails.length + htmlWinDetails.length) < 2) {
       const answer = await showMessageYesNo("There's only 1 open window. Are you sure you want to continue to make a Window Set?")
       if (answer === 'No') {
         return
       }
     }
-    const htmlWinDetails: Array<HTMLWinDetails> = NotePlan.htmlWindows.map((win) => {
-      const winRect = win.windowRect
-      return {
-        id: win.id,
-        customId: win.customId,
-        // title: win.title,
-        type: win.type,
-        // filename: win.filename
-        x: winRect.x,
-        y: winRect.y,
-        width: winRect.width,
-        height: winRect.height,
-      }
-    })
 
     // Get current saved set names
-    const savedWindowSets = DataStore.preference('windowSets')
+    const savedWindowSets: Array<WindowSet> = readWindowSetDefinitions(config.folderForDefinitions) ?? []
     let choice = 0
     let setName = ''
     let isNewSet = false
@@ -411,19 +432,16 @@ export async function saveWindowSet(): Promise<void> {
 
     // Offer current set names and/or offer to create new one
     if (savedWindowSets) {
-      clo(savedWindowSets, 'savedWindowSets')
-      windowSets = Array(savedWindowSets ?? [])
-      logDebug('saveWindowSet', `found ${String(windowSets.length)} windowSets`)
-
+      logDebug('saveWindowSet', `found ${String(savedWindowSets.length)} windowSets`)
       const nameOptions: Array<Object> = []
       nameOptions.push({ value: 0, label: "+ New window set" })
-      for (let i = 0; i < windowSets.length; i++) {
-        const thisWindowSet = windowSets[i]
+      for (let i = 0; i < savedWindowSets.length; i++) {
+        const thisWindowSet = savedWindowSets[i]
         nameOptions.push({ value: i + 1, label: thisWindowSet.name })
       }
       const res = await chooseOption('Enter new window set name, or which one to update', nameOptions, 0)
-      if (!res) {
-        logInfo('saveWindowSet', `User cancelled operation.`)
+      if (typeof res === 'boolean' && !res) {
+        logInfo('saveWindowSet', `User cancelled operation: ${res}.`)
         return
       }
       choice = res
@@ -437,6 +455,7 @@ export async function saveWindowSet(): Promise<void> {
         isNewSet = true
       } else {
         setName = nameOptions[res]
+        logDebug('saveWindowSet', `User selected existing WS '${setName}'`)
       }
     } else {
       // No current saved window sets
@@ -449,44 +468,186 @@ export async function saveWindowSet(): Promise<void> {
       isNewSet = true
     }
 
-    // TODO: Check to see if any editor windows are calendar dates
-
-    // TODO: If so, offer to make them a relative date to today/this week etc.
-
-    // TODO: Make relative calcs
-
-    // TODO: Do basics of Plugins
-
-    // Get type of window by looking at the second open window (the first always being 'main')
-    const actionType = (editorWinDetails[1].type === 'split') ? 'split' : 'floating'
-    const thisWS: WindowSet = { name: setName, editorWindows: editorWinDetails, htmlWindows: htmlWinDetails, action: actionType }
-    clo(thisWS, 'thisWS')
-
-    // Add or update this WS
-    if (isNewSet) {
-      // Add this one
-      windowSets.push(thisWS)
-    } else {
-      let c = 0
-      let found = false
-      for (const set of windowSets) {
-        clo(set, 'set')
-        if (set.name === setName) {
-          // Update this one
-          windowSets[c] = thisWS
-          found = true
-          break
-        }
-        c++
-      }
+    // Start making WS object to save
+    let thisWSToSave: WindowSet = {
+      name: setName,
+      closeOtherWindows: true,
+      editorWindows: [],
+      htmlWindows: [],
     }
-    clo(windowSets)
+
+    // First process Editor windows
+    let ewCount = 0
+    const firstWindow = editorWinDetails[0]
+    for (const ew of editorWinDetails) {
+      let tempFilename = ''
+      const thisNote = DataStore.projectNoteByFilename(ew.filename)
+      let tempTitle = displayTitle(thisNote)
+
+      // Check to see if any editor windows are calendar dates
+      if (ew.noteType === 'Calendar') {
+        // Offer to make them a relative date to today/this week etc.
+        // Turn this into a daily date at start of period
+        const thisDateStr = getISODateStringFromYYYYMMDD(getDateStringFromCalendarFilename(ew.filename))
+        // const thisDateStr = getISODateStringFromYYYYMMDD(getDateStrForStartofPeriodFromCalendarFilename(ew.filename))
+        // Get number of days between today and the calendar filename
+        logDebug('getDateStringFromCalendarFilename(ew.filename)', getDateStringFromCalendarFilename(ew.filename)) // 2023-W36
+        logDebug('thisDateStr', thisDateStr) // FIXME: error
+        // const daysDiff = daysBetween(getTodaysDateHyphenated(), thisDateStr)
+        // logDebug('daysDiff', String(daysDiff))
+
+        const relativeDateStr = `{${relativeDateCodeFromDateString(thisDateStr)}}`
+        logDebug('relativeDateStr', relativeDateStr)
+        const res = await showMessageYesNoCancel(`Open window '${thisDateStr}' is a calendar note. Do you want to make it a relative date '${relativeDateStr}'?`)
+        if (res === 'Yes') {
+          tempFilename = relativeDateStr
+          tempTitle = `relative date ${relativeDateStr}` // TODO: make proper relative name
+        } else if (res === 'No') {
+          tempFilename = ew.filename
+          // title from calendar note already set
+        } else if (res === 'Cancel') {
+          logInfo('saveWindowSet', `User cancelled operation.`)
+          return
+        }
+      } else {
+        // A project note: nothing particular to do
+      }
+    // Get type of window by looking at the second open window (the first always being 'main')
+
+      // Note: API docs say Editor.type can be main | split | floating | unsupported
+      // FIXME(Eduard): but I'm only getting Notes | Calendar
+      // const actionType = (editorWinDetails.length > 1 && editorWinDetails[1].type === 'split') ? 'split' : 'floating'
+
+      // TODO(later): remove this workaround
+      let actionType = 'floating'
+      if (ewCount === 0) {
+        actionType = 'main'
+      } else {
+        if (ew.x === firstWindow.x && ew.y === firstWindow.y && ew.width === firstWindow.width && ew.height === firstWindow.height) {
+          actionType = 'split'
+        }
+      }
+      // Save other details
+      let thisEWToSave: EditorWinDetails = {
+        noteType: ew.noteType,
+        x: ew.x,
+        y: ew.y,
+        width: ew.width,
+        height: ew.height,
+        openAction: actionType,
+        filename: tempFilename,
+      }
+      if (tempTitle) {
+        thisEWToSave.title = tempTitle
+      }
+      clo(thisWSToSave, 'thisWSToSave')
+      thisWSToSave.editorWindows.push(thisEWToSave)
+      ewCount++
+    }
+
+    // Now process HTML windows
+    // Currently works from lookup list defined at top
+    for (const thisHtmlWinDetails of htmlWinDetails) {
+      const thisWindowId = thisHtmlWinDetails.customId ?? 'n/a'
+      logDebug('saveWindowSet', `for thisHtmlWinDetails: ${thisWindowId}`)
+      const thisPWAC = pluginWindowsAndCommands.filter((p) => thisWindowId === p.pluginWindowId)
+      const thisHWPluginID = (thisPWAC)
+        ? thisHtmlWinDetails.pluginID : '? needs to be set from plugin'
+      const thisHWPluginCommand = (thisPWAC)
+        ? thisHtmlWinDetails.pluginCommand : '? needs to be set from ' + thisWindowId
+      const thisHWToSave = {
+        type: thisHtmlWinDetails.type,
+        pluginID: thisHWPluginID,
+        pluginCommand: thisHWPluginCommand,
+        customId: thisHtmlWinDetails.customId,
+        x: thisHtmlWinDetails.x,
+        y: thisHtmlWinDetails.y,
+        width: thisHtmlWinDetails.width,
+        height: thisHtmlWinDetails.height,
+      }
+      clo(thisHWToSave, 'thisHWToSave')
+      thisWSToSave.htmlWindows.push(thisHWToSave)
+    }
+
     // Save to preferences store
-    DataStore.setPreference('windowSets', windowSets)
-    logDebug('saveWindowSet', `Saved window set '${setName}'`)
-    logWindowSets()
+    // Add or update this WS
+    // if (isNewSet) {
+    //   // Add this one
+    //   windowSets.push(thisWSToSave)
+    // } else {
+    //   let c = 0
+    //   let found = false
+    //   for (const set of windowSets) {
+    //     clo(set, 'set')
+    //     if (set.name === setName) {
+    //       // Update this one
+    //       windowSets[c] = thisWSToSave
+    //       found = true
+    //       break
+    //     }
+    //     c++
+    //   }
+    // }
+    // clo(windowSets, 'windowSets')
+    // DataStore.setPreference('windowSets', windowSets)
+
+    // Save to WS note
+    // Add or update this WS
+    const WSNoteTitle = `Window Sets`
+    let res = false
+    if (isNewSet) {
+      // Prepend to WS note
+      logDebug('saveWindowSet', `About to prepend window set '${setName}' to WindowSet note ...`)
+      res = await writeWSToNote('insert', config.folderForDefinitions, WSNoteTitle, thisWSToSave)
+    } else {
+      // Replace in WS note
+      // FIXME: fails for existing note, as 'setName' undefined
+      logDebug('saveWindowSet', `About to replace window set '${setName}' in WindowSet note ...`)
+      res = await writeWSToNote('update', config.folderForDefinitions, WSNoteTitle, thisWSToSave)
+    }
+    // Report success/failure
+    if (res) {
+      logDebug('saveWindowSet', `Saved window set '${setName}' to WindowSet note`)
+    } else {
+      logError('saveWindowSet', `Failed to save window set '${setName}' to WindowSet note`)
+    }
   } catch (error) {
     logError('saveWindowSet', JSP(error))
+  }
+}
+
+/**
+ * Write the supplied WindowSet to the specified NP note, either 'insert'ing or 'update'ing
+ * @param {string} operation: 'insert' or 'update'
+ * @param {string} noteTitle to write to
+ * @param {WindowSet} thisWSToSave
+ * @returns
+ */
+export async function writeWSToNote(operation: string, noteFolder: string, noteTitle: string, thisWSToSave: WindowSet): Promise<boolean> {
+  try {
+    logDebug(pluginJson, `writeWSToNote() starting '${operation}' for '${thisWSToSave.name}' to folder '${noteFolder}' title '${noteTitle}'`)
+    const WSNote: ?TNote = await getOrMakeNote(noteTitle, noteFolder)
+    if (!WSNote) {
+      throw new Error(`writeWSToNote() no note found for '${noteTitle}' in folder '${noteFolder}'`)
+    }
+    logDebug('writeWSToNote', `- ${displayTitle(WSNote)} / ${noteTitle}`)
+
+    // Make string from WindowSet object
+    const thisWindowSetStr = JSON.stringify(thisWSToSave, null, 2)
+    // logDebug('writeWSToNote', `writeWSToNote() thisWindowSetStr:\n${thisWindowSetStr}`)
+    // Add or update to note
+    let res = false
+    if (operation === 'update') {
+      res = updateCodeBlock(WSNote, thisWindowSetStr, 'json')
+    }
+    else {
+      res = addCodeBlock(WSNote, thisWindowSetStr, 'json')
+    }
+
+    return res
+  } catch (error) {
+    logError(pluginJson, `writeWSToNote: ${error.message}`)
+    return false
   }
 }
 
@@ -610,7 +771,7 @@ export async function openWindowSet(setName: string = ''): Promise<boolean> {
       logDebug('openWindowSet', `- trying note filename '${ew.filename}' with openAction ${ew.openAction}`)
       if (ew.openAction === 'floating') {
         // Open in a full window pane
-        switch (ew.type) {
+        switch (ew.noteType) {
           case 'Calendar': {
             // if this is a relative date, calculate the actual date
             if (resourceToOpen.match(RE_OFFSET_DATE)) {
@@ -633,7 +794,7 @@ export async function openWindowSet(setName: string = ''): Promise<boolean> {
         }
       } else if (ew.openAction === 'split') {
         // Open in a split window
-        switch (ew.type) {
+        switch (ew.noteType) {
           case 'Calendar': {
             // if this is a relative date, calculate the actual date
             if (resourceToOpen.match(RE_OFFSET_DATE)) {