+++ jgclark.WindowSets/CHANGELOG.md
@@ -1,9 +1,9 @@
 # What's changed in 🖥️ Window Sets?
 _Please also see the Plugin [README](https://github.com/NotePlan/plugins/blob/main/jgclark.WindowSets/README.md)._
 
-## [0.2.0] - 2023-08-28
+## [0.2.0] - 2023-09-10 (unfinished; unreleased -- decided to change design)
 - Window Set definitions now live in notes. See README for details. Can now have any number of defined window sets.
-- ??? **/Save window set** command. Now includes window size and position for floating windows.
+- **/Save window set** command. Now includes window size and position for floating windows. (Incomplete: doesn't yet work for updating existing Window Set definition.)
 
 ## [0.1.0] - 2023-04-04
 - First version. Note: only catering for 5 window sets, due to limitiation of configuration system.